Multi_ADC Config
Configured the ioc map to include more ADCs as requested by Daniel. Clocks are the same as before. With 2 Single ended and 2 differential configurations, with DMA in circular mode and the ADCs doing continuous conversions. Currently the only issue with the project is using the CLion debugger to check the memory to see if the ADC is actually properly pulling everything in